when processing message, always reshow plot.
I previously assumed that the plot need only be shown if it wasn't
already in essgd-plot-nums.  However, as Laurent showed, if you update
a plot with a title, the plot number does not change.

The fix is simple: always update plot rather than checking to see
if plot has not yet appeared.

Closes #2